get common FN ICCs of tools
