+++ src/test/TestToolEvaluateClient.java
@@ -1,0 +1,78 @@
+package test;
+
+import main.java.MainClass;
+import main.java.MyConfig;
+import main.java.SummaryLevel;
+import org.junit.jupiter.api.Test;
+
+import java.io.File;
+
+/**
+ * @Author hanada
+ * @Date 2022/3/11 15:27
+ * @Version 1.0
+ */
+public class TestToolEvaluateClient {
+
+
+    @Test
+    public void testConfig() {
+
+        setArgs();
+        setMySwitch();
+        MainClass.startAnalyze();
+        System.out.println("Finish...\n");
+        System.exit(0);
+    }
+
+    private void setArgs() {
+        String path;
+        path = "apk/";
+        String name;
+        name = "ICCBotBench";
+        String client = "CTGClient";
+		client = "ToolEvaluateClient";
+
+        MyConfig.getInstance().setAppName(name + ".apk");
+        MyConfig.getInstance().setAppPath(path + File.separator);
+        MyConfig.getInstance().setClient(client);
+        MyConfig.getInstance().setGatorClient("GUIHierarchyPrinterClient");
+        MyConfig.getInstance().setGatorClient("ActivityTransitionAnalysisClient");
+        MyConfig.getInstance().setMaxPathNumber(30);
+        MyConfig.getInstance().setMaxFunctionExpandNumber(5); //10?
+        MyConfig.getInstance().setMaxObjectSummarySize(100);
+        MyConfig.getInstance().setResultWarpperFolder("results/" + File.separator);
+        MyConfig.getInstance().setResultFolder(MyConfig.getInstance().getResultWarpperFolder()+ "output" + File.separator);
+        MyConfig.getInstance().setTimeLimit(10);
+        MyConfig.getInstance().setAndroidJar("lib/platforms");
+
+    }
+
+    /**
+     * analyze parameters for evaluation
+     */
+    private static void setMySwitch() {
+        MyConfig.getInstance().getMySwithch().setDummyMainSwitch(false);
+        MyConfig.getInstance().getMySwithch().setCallBackSwitch(true);
+        MyConfig.getInstance().getMySwithch().setFunctionExpandSwitch(true);
+
+        MyConfig.getInstance().getMySwithch().setAsyncMethodSwitch(true);
+        MyConfig.getInstance().getMySwithch().setPolymSwitch(true);
+
+        MyConfig.getInstance().getMySwithch().setAdapterSwitch(true);
+        MyConfig.getInstance().getMySwithch().setStringOpSwitch(true);
+        MyConfig.getInstance().getMySwithch().setStaticFieldSwitch(true);
+
+        MyConfig.getInstance().getMySwithch().setFragmentSwitch(true);
+        MyConfig.getInstance().getMySwithch().setLibCodeSwitch(true);
+        MyConfig.getInstance().getMySwithch().setWrapperAPISwitch(true);
+
+        MyConfig.getInstance().getMySwithch().setImplicitLaunchSwitch(true);
+        MyConfig.getInstance().getMySwithch().setDynamicBCSwitch(true);
+
+        MyConfig.getInstance().getMySwithch().setSummaryStrategy(SummaryLevel.object);
+        MyConfig.getInstance().getMySwithch().setVfgStrategy(true);
+        MyConfig.getInstance().getMySwithch().setCgAnalyzeGroupedStrategy(false);
+    }
+
+}
